connected input/output tabs, added output pane
input/output tabs are connected with some logic to see if a request has been made, and the output tab now has a scrolleable text panel to show off data. Next step is to implement the (final?) tab, one that shows a graph for the planned averages/min/max values requests.